1.签到功能移植 2.权限功能移植 3.进一步封装 Group 和 Friend 4.色图功能移植完毕

+++ src/config/config.rs
@@ -22,7 +22,7 @@ pub struct BiliCookieConfig {
 #[derive(Debug, PartialEq, serde::Serialize, serde::Deserialize, Clone)]
 pub struct SetuConfig {
     pub local_storage: bool,
-    pub withdraw_time: i32,
+    pub withdraw_time: u64,
 }
 
 #[derive(Debug, PartialEq, serde::Serialize, serde::Deserialize, Clone)]
